Refactor exit integration test flows to helper file so they can be reused between concerns
